requirements.txt
# Core Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
llamaindex>=0.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration and Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Web Automation & Scraping
browseruse>=0.1.0
playwright>=1.40.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.25.0

# AI & ML
ollama>=0.1.0
openai>=1.3.0
anthropic>=0.7.0
sentence-transformers>=2.2.0
spacy>=3.7.0
transformers>=4.35.0

# Data & Storage
qdrant-client>=1.7.0
pinecone-client>=2.2.0
chromadb>=0.4.0
sqlite-utils>=3.35.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
loguru>=0.7.0

# APIs & Integrations
google-api-python-client>=2.110.0
linkedin-api>=2.2.0
tweepy>=4.14.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional Dependencies for specific features
# Uncomment as needed:
# streamlit>=1.28.0           # For web UI
# gradio>=4.0.0               # Alternative web UI
# celery>=5.3.0               # For background tasks
# redis>=5.0.0                # For caching and queues
